Move SqlTaskStatus to processing for better reuse.
